fix(front): address review on guild/channel context menu
