+++ frontend/src/shared/lib/use-notifications.ts
@@ -1,23 +1,26 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
-import { ApiError } from '../api/client';
 import { dispatchFriendsChanged } from './friends-events';
 import {
   buildNotificationEventsUrl,
-  deleteNotificationPreference,
   dismissNotification,
   getUnreadNotificationCount,
-  listNotificationPreferences,
   listNotifications,
   markAllNotificationsRead,
   markNotificationRead,
   NOTIFICATION_RECEIVE_EVENT,
-  setNotificationPreference,
   type ListNotificationsQuery,
-  type NotificationDto,
-  type NotificationPreferenceDto,
-  type NotificationScopeType
+  type NotificationDto
 } from '../api/notification';
 import { getAccessToken } from './session';
+import { useNotificationPrefs } from './notification-prefs-store';
+import { isSuppressedByMute } from './notification-mute';
+
+// re-exported so existing importers (notification-card) keep their import path
+export {
+  isScopeMuted,
+  getNotificationMuteScopes,
+  type NotificationMuteScope
+} from './notification-mute';
 
 const PAGE_SIZE = 50;
 const RECONNECT_BASE_DELAY_MS = 2_000;
@@ -27,54 +30,6 @@ export type NotificationFilter = {
   unreadOnly: boolean;
   includeDismissed: boolean;
 };
-
-export type NotificationMuteScope = {
-  scopeType: NotificationScopeType;
-  scopeId: string;
-};
-
-export function isScopeMuted(
-  preferences: NotificationPreferenceDto[],
-  scopeType: NotificationScopeType,
-  scopeId: string
-): boolean {
-  const now = Date.now();
-  return preferences.some(
-    (preference) =>
-      preference.scope_type === scopeType &&
-      preference.scope_id === scopeId &&
-      preference.muted &&
-      (preference.muted_until == null || Date.parse(preference.muted_until) > now)
-  );
-}
-
-// the scopes a notification can be muted under; also drives the per-item mute buttons
-export function getNotificationMuteScopes(notification: NotificationDto): NotificationMuteScope[] {
-  switch (notification.type) {
-    case 'mention':
-      return [
-        { scopeType: 'channel', scopeId: notification.payload.channel_id },
-        { scopeType: 'guild', scopeId: notification.payload.guild_id }
-      ];
-    case 'guild_invite':
-    case 'guild_welcome':
-      // source_id carries the guild id for both types
-      return notification.source_id
-        ? [{ scopeType: 'guild', scopeId: notification.source_id }]
-        : [];
-    default:
-      return [];
-  }
-}
-
-function isSuppressedByMute(
-  notification: NotificationDto,
-  preferences: NotificationPreferenceDto[]
-): boolean {
-  return getNotificationMuteScopes(notification).some(({ scopeType, scopeId }) =>
-    isScopeMuted(preferences, scopeType, scopeId)
-  );
-}
 
 function buildListQuery(filter: NotificationFilter): ListNotificationsQuery {
   return {
@@ -96,7 +51,9 @@ export function useNotifications() {
     unreadOnly: true,
     includeDismissed: false
   });
-  const [preferences, setPreferences] = useState<NotificationPreferenceDto[]>([]);
+  // preferences (and their mutations) are owned by the shared store so the
+  // notification panel, the sidebar dimming and the context menus stay in sync
+  const { preferences, mute, unmute } = useNotificationPrefs();
 
   // bumped on every feed fetch; responses that come back after a newer fetch
   // started are discarded instead of clobbering the newer results
@@ -138,20 +95,6 @@ export function useNotifications() {
   useEffect(() => {
     void fetchFeed();
   }, [fetchFeed]);
-
-  useEffect(() => {
-    let active = true;
-    listNotificationPreferences()
-      .then((items) => {
-        if (active) setPreferences(items);
-      })
-      .catch(() => {
-        // best-effort: without preferences the server still filters muted inserts
-      });
-    return () => {
-      active = false;
-    };
-  }, []);
 
   const loadMore = useCallback(async () => {
     const oldest = notifications[notifications.length - 1];
@@ -375,38 +318,6 @@ export function useNotifications() {
     },
     [notifications, filter.includeDismissed, fetchFeed]
   );
-
-  const mute = useCallback(
-    async (scopeType: NotificationScopeType, scopeId: string, mutedUntil?: string | null) => {
-      const updated = await setNotificationPreference(scopeType, scopeId, {
-        muted: true,
-        muted_until: mutedUntil ?? null
-      });
-      setPreferences((current) => [
-        ...current.filter(
-          (preference) => !(preference.scope_type === scopeType && preference.scope_id === scopeId)
-        ),
-        updated
-      ]);
-    },
-    []
-  );
-
-  const unmute = useCallback(async (scopeType: NotificationScopeType, scopeId: string) => {
-    try {
-      await deleteNotificationPreference(scopeType, scopeId);
-    } catch (err) {
-      // 404 means no preference row, which is already the unmuted default
-      if (!(err instanceof ApiError && err.status === 404)) {
-        throw err;
-      }
-    }
-    setPreferences((current) =>
-      current.filter(
-        (preference) => !(preference.scope_type === scopeType && preference.scope_id === scopeId)
-      )
-    );
-  }, []);
 
   return {
     notifications,
